Adapt gladys-gateway-js to gladys 4

+++ src/gladys-gateway-js/lib/helpers.js
@@ -1,10 +1,10 @@
 function ab2str(buf) {
-  return String.fromCharCode.apply(null, new Uint16Array(buf));
+  return String.fromCharCode.apply(null, new Uint8Array(buf));
 }
 
 function str2ab(str) {
   let buf = new ArrayBuffer(str.length * 2);
-  let bufView = new Uint16Array(buf);
+  let bufView = new Uint8Array(buf);
   for (let i = 0, strLen = str.length; i < strLen; i++) {
     bufView[i] = str.charCodeAt(i);
   }
@@ -18,7 +18,7 @@ function appendBuffer(buffer1, buffer2) {
   return tmp.buffer;
 }
 
-function sanitizePassPhrase(passphrase) {
+function sanitizePassPhrase(passphrase) {
   return passphrase.trim().normalize('NFKD');
 }
